add repo to the app
